fix: move more modules to langbot packages
